+++ config/rustdesk.go
@@ -18,7 +18,7 @@ type Rustdesk struct {
 	Key             string `mapstructure:"key"`
 	KeyFile         string `mapstructure:"key-file"`
 	Personal        int    `mapstructure:"personal"`
-	//webclient-magic-queryonline
+	// webclient-magic-queryonline
 	WebclientMagicQueryonline int    `mapstructure:"webclient-magic-queryonline"`
 	WsHost                    string `mapstructure:"ws-host"`
 }