+++ lib/jwt/jwt.go
@@ -2,8 +2,9 @@ package jwt
 
 import (
 	"fmt"
-	"github.com/golang-jwt/jwt/v5"
 	"time"
+
+	"github.com/golang-jwt/jwt/v5"
 )
 
 type Jwt struct {
@@ -44,7 +45,7 @@ func (s *Jwt) GenerateToken(userId uint) string {
 }
 
 func (s *Jwt) ParseToken(tokenString string) (uint, error) {
-	token, err := jwt.ParseWithClaims(tokenString, &UserClaims{}, func(token *jwt.Token) (interface{}, error) {
+	token, err := jwt.ParseWithClaims(tokenString, &UserClaims{}, func(_ *jwt.Token) (interface{}, error) {
 		return s.Key, nil
 	})
 	if err != nil {