+++ lib/orm/postgresql.go
@@ -2,10 +2,11 @@ package orm
 
 import (
 	"fmt"
+	"time"
+
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
 	"gorm.io/gorm/logger"
-	"time"
 )
 
 type PostgresqlConfig struct {