+++ model/custom_types/auto_json.go
@@ -3,7 +3,6 @@ package custom_types
 import (
 	"database/sql/driver"
 	"encoding/json"
-	"errors"
 	"fmt"
 )
 
@@ -19,21 +18,21 @@ func (j *AutoJson) Scan(value interface{}) error {
 	case string:
 		strValue = v
 	default:
-		return errors.New(fmt.Sprintf("Failed Scan AutoJson value: %v", value))
+		return fmt.Errorf("failed Scan AutoJson value: %v", value)
 	}
 	bytes := []byte(strValue)
-	//bytes, ok := value.([]byte)
-	//if !ok {
+	// bytes, ok := value.([]byte)
+	// if !ok {
 	//	return errors.New(fmt.Sprint("Failed Scan AutoJson value:", value))
 	//}
 
-	if bytes == nil || len(bytes) == 0 {
+	if len(bytes) == 0 {
 		*j = AutoJson(json.RawMessage{'[', ']'})
 		return nil
 	}
 	result := &json.RawMessage{}
 	err := json.Unmarshal(bytes, result)
-	//解析json错误 返回空
+	// 解析json错误 返回空
 	if err != nil {
 		*j = AutoJson(json.RawMessage{'[', ']'})
 		return nil